refactor: divide StockCorrectionBody in 3 components
